+++ pkg/vm/engine/tae/containers/types.go
@@ -6,12 +6,12 @@ import (
 	"github.com/RoaringBitmap/roaring"
 	"github.com/RoaringBitmap/roaring/roaring64"
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/stl"
-	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/stl/container"
+	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/stl/containers"
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/types"
 )
 
 type MemAllocator = stl.MemAllocator
-type Options = container.Options
+type Options = containers.Options
 type Bytes = stl.Bytes
 
 type Vector interface {
