+++ pkg/vm/engine/tae/stl/containers/strvec.go
@@ -1,4 +1,4 @@
-package container
+package containers
 
 import (
 	"fmt"
