+++ pkg/vm/engine/tae/stl/containers/types.go
@@ -1,4 +1,4 @@
-package container
+package containers
 
 import (
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/stl"
