+++ pkg/vm/engine/tae/stl/containers/vector_test.go
@@ -1,4 +1,4 @@
-package container
+package containers
 
 import (
 	"bytes"
